Add a non-interactive -y flag to our installer.
This is more reliable than advising people to  for all the same reasons I just removed
that from this script.

This supersedes the ASK_REPORT_QUESTION option, assuming No by default.
In the future we can add a non-interactive flag for that, and all the other interactive questions,
if it really seems important ((but our Sentry account has expired so we're not even using it right now)).